feat: add monitoring stack
Every cluster now has the podMonitor enabled and the script
scripts/setup_monitoring.sh was created to deploy Prometheus Operator
and Grafana Operator, using these operators we create:

* Prometheus instance
* Grafana Deployment
* Grafana Datasource (connecting to Prometheus instance)
* Grafana Dashboard (from the project cloudnative-pg/grafana-dashboard)

After the cluster are updated the Grafana dashboard can be accessed
redirecting the 3000 port of the Grafana deployment (namespace grafana) on
any of the two kind clusters.

The dashboard is loaded from the URL of the json file in the repository,
this means that if the content is changed also the dashboard will be updated.

Signed-off-by: Jonathan Gonzalez V. <jonathan.abdiel@gmail.com>